# user/cmd/api/desc/rest.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "",
    "version": ""
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/v1/createUser": {
      "post": {
        "summary": "创建用户",
        "operationId": "createUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateUserResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserReq"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/v1/deleteUser": {
      "post": {
        "summary": "根据id删除用户",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/DeleteUserResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteUserReq"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/v1/getUserById": {
      "post": {
        "summary": "根据id获取用户",
        "operationId": "getUserById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetUserByIdResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetUserByIdReq"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/v1/getUsers": {
      "post": {
        "summary": "获取用户列表",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetUserListResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetUserListReq"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/user/v1/loginUser": {
      "post": {
        "summary": "单点用户登录",
        "operationId": "loginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/LoginUserResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUserReq"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    }
  },
  "definitions": {
    "CreateUserReq": {
      "type": "object",
      "properties": {
        "client_id": {
          "type": "integer",
          "format": "int32"
        },
        "system_id": {
          "type": "integer",
          "format": "int32"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "CreateUserReq",
      "required": [
        "client_id",
        "system_id",
        "phone",
        "password"
      ]
    },
    "CreateUserResp": {
      "type": "object",
      "title": "CreateUserResp"
    },
    "DeleteUserReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "DeleteUserReq",
      "required": [
        "id"
      ]
    },
    "DeleteUserResp": {
      "type": "object",
      "title": "DeleteUserResp"
    },
    "GetUserByIdReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "GetUserByIdReq",
      "required": [
        "id"
      ]
    },
    "GetUserByIdResp": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "title": "GetUserByIdResp",
      "required": [
        "user"
      ]
    },
    "GetUserListReq": {
      "type": "object",
      "title": "GetUserListReq"
    },
    "GetUserListResp": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      },
      "title": "GetUserListResp",
      "required": [
        "users"
      ]
    },
    "LoginUserReq": {
      "type": "object",
      "properties": {
        "client_id": {
          "type": "integer",
          "format": "int32"
        },
        "phone": {
          "type": "string"
        },
        "system_id": {
          "type": "integer",
          "format": "int32"
        },
        "verificationCode": {
          "type": "string"
        }
      },
      "title": "LoginUserReq",
      "required": [
        "client_id",
        "phone",
        "system_id",
        "verificationCode"
      ]
    },
    "LoginUserResp": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "title": "LoginUserResp",
      "required": [
        "user"
      ]
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "create_time": {
          "type": "string"
        },
        "update_time": {
          "type": "string"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32"
        },
        "account": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "gender": {
          "type": "integer",
          "format": "int32"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "id_card": {
          "type": "string"
        },
        "source": {
          "type": "integer",
          "format": "int32"
        },
        "system_id": {
          "type": "integer",
          "format": "int32"
        },
        "client_id": {
          "type": "integer",
          "format": "int32"
        },
        "union_id": {
          "type": "string"
        },
        "is_enable": {
          "type": "integer",
          "format": "int32"
        },
        "last_login_time": {
          "type": "string"
        },
        "create_user": {
          "type": "integer",
          "format": "int64"
        },
        "update_user": {
          "type": "integer",
          "format": "int64"
        },
        "token": {
          "type": "string"
        },
        "session_id": {
          "type": "string"
        },
        "token_expire_time": {
          "type": "string"
        }
      },
      "title": "User",
      "required": [
        "id",
        "create_time",
        "update_time",
        "isDeleted",
        "account",
        "avatar",
        "gender",
        "phone",
        "password",
        "id_card",
        "source",
        "system_id",
        "client_id",
        "union_id",
        "is_enable",
        "last_login_time",
        "create_user",
        "update_user",
        "token",
        "session_id",
        "token_expire_time"
      ]
    }
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Enter JWT Bearer token **_only_**",
      "name": "Authorization",
      "in": "header"
    }
  }
}